Refactor code in app.js
